real chromosome instead of integer

+++ src/main/java/fitness/ForTestingFitnessFunction.java
@@ -6,7 +6,7 @@ import entities.Person;
 public class ForTestingFitnessFunction implements FitnessFunction {
 
     public double apply(Person person) {
-        double fitness = (1 / (double) person.getChromosomes().stream().mapToInt(Chromosome::getValue).sum());
+        double fitness = (1 / person.getChromosomes().stream().mapToDouble(Chromosome::getValue).sum());
         person.setFitnessValue(fitness);
         return fitness;
     }
